Move get_available_genres to utils
